Create maze_extra_credit.txt
An extra maze given to be read in.  This was for an extra credit portion, the difference being there are not a set number of "rooms" between fast track entrances.  This requires the robot to learn and retain knowledge about the path.  This maze will also work with the model.